[refactor] 페이지 구분을 위한 하위폴더 생성 및 파일 이동
