+++ src/components/mylocationstore/KakaoMap.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
 import { useEffect, useState } from "react";
-import { Store } from "../../page/MyLocationStore";
+import { Store } from "../../page/mylocationstore/MyLocationStore";
 import { ROUTER_PATH } from "../../constants/constants";
 
 declare global {
