+++ src/page/auth/Signup.tsx
@@ -1,9 +1,9 @@
 import React from "react";
-import OuterLayout from "../components/common/OuterLayout";
-import Footer from "../components/common/Footer";
-import Nav from "../components/common/Nav";
-import PageTitle from "../components/common/PageTitle";
-import SignUpPage from "../components/signup/SignUpPage";
+import OuterLayout from "../../components/common/OuterLayout";
+import Footer from "../../components/common/Footer";
+import Nav from "../../components/common/Nav";
+import PageTitle from "../../components/common/PageTitle";
+import SignUpPage from "../../components/signup/SignUpPage";
 
 const Signup: React.FC = () => {
   return (
