+++ src/page/mylocationstore/MyLocationStore.tsx
@@ -2,15 +2,15 @@ import React from "react";
 
 import { useState, useEffect } from "react";
 
-import OuterLayout from "../components/common/OuterLayout";
-import PageTitle from "../components/common/PageTitle";
-import Footer from "../components/common/Footer";
-import Nav from "../components/common/Nav";
-import KakaoMap from "../components/mylocationstore/KakaoMap";
-import StoreList from "../components/mylocationstore/StoreList";
-import StyleSlider from "../components/mylocationstore/StyleSlider";
-import useAxios from "../hooks/useAxios";
-import useToastMessage from "../hooks/useToastMessage";
+import OuterLayout from "../../components/common/OuterLayout";
+import PageTitle from "../../components/common/PageTitle";
+import Footer from "../../components/common/Footer";
+import Nav from "../../components/common/Nav";
+import KakaoMap from "../../components/mylocationstore/KakaoMap";
+import StoreList from "../../components/mylocationstore/StoreList";
+import StyleSlider from "../../components/mylocationstore/StyleSlider";
+import useAxios from "../../hooks/useAxios";
+import useToastMessage from "../../hooks/useToastMessage";
 
 export interface Store {
   storeNo: number;
