+++ src/page/mypage/MyReservation.tsx
@@ -2,14 +2,14 @@ import React, { useEffect, useState } from "react";
 import { format, parse, parseISO } from "date-fns";
 import { ko } from "date-fns/locale";
 
-import PageNothing from "../components/common/PageNothing";
-import ModalDelete from "../components/common/ModalDelete";
-import OuterLayout from "../components/common/OuterLayout";
-import PageTitle from "../components/common/PageTitle";
-import Nav from "../components/common/Nav";
-import MyReservationCard from "../components/myreservation/MyReservationCard";
-import useAxios from "../hooks/useAxios";
-import useToastMessage from "../hooks/useToastMessage";
+import PageNothing from "../../components/common/PageNothing";
+import ModalDelete from "../../components/common/ModalDelete";
+import OuterLayout from "../../components/common/OuterLayout";
+import PageTitle from "../../components/common/PageTitle";
+import Nav from "../../components/common/Nav";
+import MyReservationCard from "../../components/myreservation/MyReservationCard";
+import useAxios from "../../hooks/useAxios";
+import useToastMessage from "../../hooks/useToastMessage";
 
 interface MyReservationDatasType {
   reservationNo: number;
