+++ src/page/mypage/RegisterQnA.tsx
@@ -1,11 +1,11 @@
 import React, { useState } from "react";
 
-import RegisterQnAForm from "../components/registerqna/RegisterQnAForm";
-import PageComplete from "../components/common/PageComplete";
-import OuterLayout from "../components/common/OuterLayout";
-import PageTitle from "../components/common/PageTitle";
-import Nav from "../components/common/Nav";
-import { ROUTER_PATH } from "../constants/constants";
+import RegisterQnAForm from "../../components/registerqna/RegisterQnAForm";
+import PageComplete from "../../components/common/PageComplete";
+import OuterLayout from "../../components/common/OuterLayout";
+import PageTitle from "../../components/common/PageTitle";
+import Nav from "../../components/common/Nav";
+import { ROUTER_PATH } from "../../constants/constants";
 
 const RegisterQnA: React.FC = () => {
   const [step, setStep] = useState(1);
